docs: implement saving language layout after reload the page
